+++ src/changelog.md
@@ -1,8 +1,14 @@
 # Changelog
  
+## 2022-06-30
+
+* [RELEASE DEV] Released as Dev
+* [Fix] Fix number input in locale format does not validate
+
 ## 2022-05-11
 
-* [Fix] Fix percentages without decimals include in the print a dot '.' (ex.: 30.%)
+* [RELEASE DEV] Released as Dev
+* [Fix] Fix percentages without decimals include a dot '.' in the print (ex.: 30.%)
 
 ## 2022-03-31